Introduce WebFormModule and keep AuraInputModule as alias
Add Ray\WebFormModule\WebFormModule whose name lines up with the
package (ray/web-form-module) and namespace (Ray\WebFormModule), so
the public module class no longer leaks the underlying Aura.Input
implementation detail.

Ray\WebFormModule\AuraInputModule is preserved as a thin subclass of
WebFormModule so existing applications that install
'new AuraInputModule()' continue to work unchanged. The interceptor
test keeps using AuraInputModule to lock in that compatibility.

Update README, README.JA, the CSRF demo, and the fake module used in
tests to install the new WebFormModule. Add CHANGELOG entry for 1.0.1.
