+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2026-05-18
+
+### Added
+
+- `Ray\WebFormModule\WebFormModule` — module class whose name matches the
+  package and namespace. Use this in new code.
+
+### Deprecated
+
+- `Ray\WebFormModule\AuraInputModule` is now a thin subclass of
+  `WebFormModule` kept for backwards compatibility. Existing applications
+  that install `new AuraInputModule()` continue to work without changes,
+  but should migrate to `new WebFormModule()`.
+
 ## [1.0.0] - 2026-05-17
 
 ### Changed
@@ -75,5 +89,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 See git history for changes prior to 1.0.0.
 
+[1.0.1]: https://github.com/ray-di/Ray.WebFormModule/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/ray-di/Ray.WebFormModule/compare/0.6.0...1.0.0
 [0.6.0]: https://github.com/ray-di/Ray.WebFormModule/releases/tag/0.6.0
