+++ src/AuraInputModule.php
@@ -4,43 +4,13 @@ declare(strict_types=1);
 
 namespace Ray\WebFormModule;
 
-use Aura\Filter\FilterFactory;
-use Aura\Html\HelperLocatorFactory;
-use Aura\Input\AntiCsrfInterface;
-use Aura\Input\Builder;
-use Aura\Input\BuilderInterface;
-use Aura\Input\Filter;
-use Aura\Input\FilterInterface;
-use Ray\AuraSessionModule\AuraSessionModule;
-use Ray\Di\AbstractModule;
-use Ray\Di\Scope;
-use Ray\WebFormModule\Annotation\FormValidation;
-use Ray\WebFormModule\Annotation\InputValidation;
-
-/** @SuppressWarnings(PHPMD.CouplingBetweenObjects) */
-class AuraInputModule extends AbstractModule
+/**
+ * Backwards-compatible alias for {@see WebFormModule}.
+ *
+ * Use {@see WebFormModule} in new code. This class is kept so existing
+ * applications that install `new AuraInputModule()` continue to work
+ * without changes.
+ */
+class AuraInputModule extends WebFormModule
 {
-    /** {@inheritDoc} */
-    protected function configure()
-    {
-        $this->install(new AuraSessionModule());
-        $this->bind(BuilderInterface::class)->to(Builder::class);
-        $this->bind(FilterInterface::class)->to(Filter::class);
-        $this->bind(AntiCsrfInterface::class)->to(AntiCsrf::class)->in(Scope::SINGLETON);
-        $this->bind(FailureHandlerInterface::class)->to(OnFailureMethodHandler::class);
-        $this->bind(FailureHandlerInterface::class)
-            ->annotatedWith('vnd_error')->to(VndErrorHandler::class)->in(Scope::SINGLETON);
-        $this->bind(HelperLocatorFactory::class);
-        $this->bind(FilterFactory::class);
-        $this->bindInterceptor(
-            $this->matcher->any(),
-            $this->matcher->annotatedWith(InputValidation::class),
-            [InputValidationInterceptor::class],
-        );
-        $this->bindInterceptor(
-            $this->matcher->any(),
-            $this->matcher->annotatedWith(FormValidation::class),
-            [AuraInputInterceptor::class],
-        );
-    }
 }
