+++ src/WebFormModule.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ray\WebFormModule;
+
+use Aura\Filter\FilterFactory;
+use Aura\Html\HelperLocatorFactory;
+use Aura\Input\AntiCsrfInterface;
+use Aura\Input\Builder;
+use Aura\Input\BuilderInterface;
+use Aura\Input\Filter;
+use Aura\Input\FilterInterface;
+use Ray\AuraSessionModule\AuraSessionModule;
+use Ray\Di\AbstractModule;
+use Ray\Di\Scope;
+use Ray\WebFormModule\Annotation\FormValidation;
+use Ray\WebFormModule\Annotation\InputValidation;
+
+/** @SuppressWarnings(PHPMD.CouplingBetweenObjects) */
+class WebFormModule extends AbstractModule
+{
+    /** {@inheritDoc} */
+    protected function configure()
+    {
+        $this->install(new AuraSessionModule());
+        $this->bind(BuilderInterface::class)->to(Builder::class);
+        $this->bind(FilterInterface::class)->to(Filter::class);
+        $this->bind(AntiCsrfInterface::class)->to(AntiCsrf::class)->in(Scope::SINGLETON);
+        $this->bind(FailureHandlerInterface::class)->to(OnFailureMethodHandler::class);
+        $this->bind(FailureHandlerInterface::class)
+            ->annotatedWith('vnd_error')->to(VndErrorHandler::class)->in(Scope::SINGLETON);
+        $this->bind(HelperLocatorFactory::class);
+        $this->bind(FilterFactory::class);
+        $this->bindInterceptor(
+            $this->matcher->any(),
+            $this->matcher->annotatedWith(InputValidation::class),
+            [InputValidationInterceptor::class],
+        );
+        $this->bindInterceptor(
+            $this->matcher->any(),
+            $this->matcher->annotatedWith(FormValidation::class),
+            [AuraInputInterceptor::class],
+        );
+    }
+}
