+++ tests/Fake/FakeModule.php
@@ -10,7 +10,7 @@ class FakeModule extends AbstractModule
     protected function configure()
     {
         $this->bind(Phpfunc::class)->to(FakePhpFunc::class);
-        $this->install(new AuraInputModule);
+        $this->install(new WebFormModule());
         $this->bind(FormInterface::class)->annotatedWith('contact_form')->to(FakeForm::class);
     }
 }
